requirements311.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.3
idna==3.4
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==6.0.4
mysql-connector-python==8.0.27
numpy==1.24.2
openai==0.27.4
pandas==2.0.0
protobuf==3.19.1
psycopg2==2.9.5
pymongo==3.11.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2023.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.3.24
tqdm==4.64.1
tzdata==2023.3
urllib3==1.26.15
Werkzeug==1.0.1
yarl==1.8.2
zipp==3.6.0
